ocr-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
easyocr==1.7.1
Pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
python-multipart==0.0.9
pillow-heif==0.18.0
